+++ tests/testthat/test_data.R
@@ -1,0 +1,5 @@
+
+test_that("Data is loaded", {
+  expect_s3_class(malaria_df1,"data.frame")
+  expect_s3_class(malaria_df2,"data.frame")
+})