+++ tests/testthat/test_get_latent_model.R
@@ -1,0 +1,4 @@
+test_that("get_latent_model workds",{
+  xx <- get_latent_model()
+  expect_type(xx, "character")
+})